fix(Mobile): recreated the /services route service
